requirements.txt
# Core Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
pymongo==4.6.1
dnspython==2.4.2

# Production Server
gunicorn==21.2.0

# Environment & Configuration
python-dotenv==1.0.0

# CORS Support
flask-cors==4.0.0
